Initial 2D and 3D simultaneous testing
Wow. I did a lot in one day.

Created a basic folder structure with debug folders in each sub folder. Put anything in here that is not relevant to the final game but can be used to test other features. If you want to know what any of the folders are for, just ping @AdoHTQ in the discord.

Installed the Universal Render Pipeline package and created a URP Asset in order to allow better control over the graphic style of the game (including the 2D and 3D) as well as creating 2 URP Renderers for the 2D and 3D modes which are set on the camera objects.

Removed the Unity sample scene. :(

Created a testing scene with a button to switch between 2D and 3D as well as some sample debug areas for each.

Added a CameraChange script that can be added to a button to toggle on and off the 2D and 3D cameras.

Created a white pixel texture which can be scaled and recolored for any solid color for debug.

URP Global Settings asset was auto generated by URP.

I sure do love day 1 of game dev. I can't wait for all my motivation to vanish. 👍